Allow modules to register their own widgets with the tkinter event binding system.
The way it's done is to look for bindable_widgets in each moduleinstance
when mods are being loaded. If there are bindable widgets, we add them to
our dictionary. Bindable widgets should implement bindevent or bindprotocol,
and this function is called when a event/protocol is requested.

Only after all mods are loaded do we start binding. Now, each bindable
widget has its own widget name (a string), so all other mods have to do now
is to request the widget name and the tkinter event as a tuple, instead of
waiting for mods to load and trying to register the events themselves
(messy!). Hooks still use the old format, but I'm considering changing hooks
to a tuple format as well.